Use local tool for EF instead of global (#52)
Setup instructions require us to use the dotnet-ef tool globally. This wasn't a problem pre-3.0 since the EF version was tied to the SDK version in `global.json`. It's not anymore and different developer machines could end up with tool versions eventually.

The tool should be installed locally for the project instead, so that there's a guarantee that every developer ends up using the same local tool version.